+++ DrillingRig.ConfigApp/AppControl/EngineSettingsSpace/EngineSettingsPartWritable.cs
@@ -3,9 +3,9 @@
 		public ushort? Inom { get; set; }
 		public ushort? Nnom { get; set; }
 		public ushort? Nmax { get; set; }
-		public double? Pnom { get; set; }
-		public double? CosFi { get; set; }
-		public double? Eff { get; set; }
+		public decimal? Pnom { get; set; }
+		public decimal? CosFi { get; set; }
+		public decimal? Eff { get; set; }
 		public ushort? Mass { get; set; }
 		public ushort? MmM { get; set; }
 		public ushort? Height { get; set; }
